make create/edit atomic transaction as well as dont save empty ingredients
