+++ recipes/recipe_create.py
@@ -1,0 +1,60 @@
+from django.db import transaction
+from rest_framework import status
+from rest_framework.parsers import MultiPartParser, FormParser
+from rest_framework.views import APIView
+import json
+from .serializers import RecipeSerializer, RecipeStepSerializer, RecipeCategorySerializer, IngredientSerializer
+from .models import Ingredient, Recipe, RecipeCategory, RecipeStep
+from rest_framework.response import Response
+from .helpers.recipe_helpers import clean_instructions, clean_ingredients
+
+class RecipeCreate(APIView):
+    parser_classes = [MultiPartParser, FormParser]
+
+    @transaction.atomic
+    def post(self, request, format=None):
+        instructions_list = json.loads(request.data['instructions_list'])
+        ingredients_list = json.loads(request.data['ingredients_list'])
+        recipe_category = request.data['recipe_category_name']
+        cleaned_instructions: list(str) = clean_instructions(instructions_list)
+        cleaned_ingredients: list(object) = clean_ingredients(ingredients_list)
+
+        serializer = RecipeSerializer(data=request.data)
+        if serializer.is_valid():
+            # saving the recipe to get it to serialize the FileUpload object for the image
+            recipe = serializer.save()
+            recipe_category, created = RecipeCategory.objects.get_or_create(category_name=recipe_category, user=recipe.user)
+
+            recipe = Recipe.objects.get(pk=recipe.id)
+            recipe.recipe_category = recipe_category
+            if recipe.image and str(recipe.image) != '':
+                recipe.image = 'https://res.cloudinary.com/djtd4wqoc/image/upload/v1643515599/' + str(recipe.image)
+            recipe.save()
+            for step_index, instructions in enumerate(cleaned_instructions):
+                step_number = step_index
+                recipe_step = recipe.steps.create(step_number=step_number,
+                                                  instructions=instructions,
+                                                  image=None, recipe=recipe)
+            for ingredients in cleaned_ingredients:
+                parsed_ingredient_name = ingredients['ingredient_name'].lower().strip()
+                parsed_quantity_unit = ingredients['quantity_unit'].lower().strip()
+                if (len(parsed_quantity_unit) > 1 and parsed_quantity_unit[-1] == 's'):
+                    parsed_quantity_unit = parsed_quantity_unit[:-1]
+                ingredient = recipe.ingredients.create(ingredient_name=parsed_ingredient_name,
+                                                       ingredient_quantity=str(
+                                                           round(float(ingredients['ingredient_quantity']), 2)),
+                                                       quantity_unit=parsed_quantity_unit, recipe=recipe)
+
+            recipe_serializer = RecipeSerializer(recipe, many=False)
+            recipe_category_serializer = RecipeCategorySerializer(recipe_category, many=False)
+            instructions_serializer = RecipeStepSerializer(RecipeStep.objects.filter(recipe=recipe), many=True)
+            ingredients_serializer = IngredientSerializer(Ingredient.objects.filter(recipe=recipe), many=True)
+            obj = {
+                'recipe': recipe_serializer.data,
+                'category': recipe_category_serializer.data,
+                'instructions': instructions_serializer.data,
+                'ingredients': ingredients_serializer.data,
+            }
+            return Response(obj)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
