+++ recipes/views.py
@@ -1,18 +1,17 @@
 
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
+from django.db import transaction
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
+
+from .helpers.recipe_helpers import clean_instructions, clean_ingredients
 from .models import Ingredient, Recipe, RecipeCategory, RecipeStep
 import json
 from .serializers import RecipeSerializer, RecipeStepSerializer, RecipeCategorySerializer, IngredientSerializer
 import requests
 from dotenv import load_dotenv
 import os
-from django.shortcuts import redirect
-from rest_framework import status
-from rest_framework.parsers import MultiPartParser, FormParser
-from rest_framework.views import APIView
 import cloudinary.api
 from django.conf import settings
 
@@ -131,55 +130,7 @@ def recipe_show(request, id):
     print('response', obj)
     return Response(obj)
 
-
-class RecipeCreate(APIView):
-    parser_classes = [MultiPartParser, FormParser]
-    def post(self, request, format=None):
-        instructions_list = json.loads(request.data['instructions_list'])
-        ingredients_list = json.loads(request.data['ingredients_list'])
-        recipe_category = request.data['recipe_category_name']
-
-        serializer = RecipeSerializer(data=request.data)
-        if(serializer.is_valid()):
-            # saving the recipe to get it to serialize the FileUpload object for the image
-            recipe = serializer.save()
-            recipe_category, created = RecipeCategory.objects.get_or_create(category_name=recipe_category, user=recipe.user)
-
-            recipe = Recipe.objects.get(pk=recipe.id)
-            recipe.recipe_category = recipe_category
-            if recipe.image and str(recipe.image) != '':
-                recipe.image = 'https://res.cloudinary.com/djtd4wqoc/image/upload/v1643515599/' + str(recipe.image)
-            recipe.save()
-            for instructions in instructions_list:
-                recipe_step = recipe.steps.create(step_number=instructions['step_number'],
-                                                  instructions=instructions['instructions'], image=None, recipe=recipe)
-
-            for ingredients in ingredients_list:
-                parsed_ingredient_name = ingredients['ingredient_name'].lower().strip()
-                parsed_quantity_unit = ingredients['quantity_unit'].lower().strip()
-                if (len(parsed_quantity_unit) > 1 and parsed_quantity_unit[-1] == 's'):
-                    parsed_quantity_unit = parsed_quantity_unit[:-1]
-                ingredient = recipe.ingredients.create(ingredient_name=parsed_ingredient_name,
-                                                       ingredient_quantity=str(
-                                                           round(float(ingredients['ingredient_quantity']), 2)),
-                                                       quantity_unit=parsed_quantity_unit, recipe=recipe)
-
-            recipe_serializer = RecipeSerializer(recipe, many=False)
-            recipe_category_serializer = RecipeCategorySerializer(recipe_category, many=False)
-            instructions_serializer = RecipeStepSerializer(RecipeStep.objects.filter(recipe=recipe), many=True)
-            ingredients_serializer = IngredientSerializer(Ingredient.objects.filter(recipe=recipe), many=True)
-            obj = {
-                'recipe': recipe_serializer.data,
-                'category': recipe_category_serializer.data,
-                'instructions': instructions_serializer.data,
-                'ingredients': ingredients_serializer.data,
-            }
-            return Response(obj)
-        else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-
+@transaction.atomic
 @api_view(['POST'])
 def recipe_search_new(request):
     user_id = request.data['user_id']
@@ -240,8 +191,9 @@ def recipe_search_new(request):
     # print(obj)
     # data=json.dumps(obj)
     return Response(obj)
-    
 
+
+@transaction.atomic
 @api_view(['POST'])
 def recipe_edit(request, id):
     user_id = request.data['user_id']
@@ -251,6 +203,8 @@ def recipe_edit(request, id):
     ingredients_list = json.loads(request.data['ingredients_list'])
     image_file = request.data['image']
     user = User.objects.get(pk=user_id)
+    cleaned_instructions: list(str) = clean_instructions(instructions_list)
+    cleaned_ingredients: list(object) = clean_ingredients(ingredients_list)
 
     recipe = Recipe.objects.get(pk=id)
     recipe_category, created = RecipeCategory.objects.get_or_create(user=user, category_name=recipe_category)
@@ -265,23 +219,20 @@ def recipe_edit(request, id):
         recipe.image = cloudinary_upload_string
 
     recipe.save()
-
     recipe.steps.all().delete()
-    for instructions in instructions_list:
-        recipe_step = recipe.steps.create(step_number=instructions['step_number'],
-                        instructions=instructions['instructions'], image=None, recipe=recipe)
+    for step_index, instructions in enumerate(cleaned_instructions):
+        step_number = step_index + 1
+        recipe_step = recipe.steps.create(step_number=step_number,
+                        instructions=instructions, image=None, recipe=recipe)
         recipe_step.save()
     recipe.ingredients.all().delete()
-    for ingredients in ingredients_list:
-        print('ingredient unit before parse', ingredients['quantity_unit'])
+    for ingredients in cleaned_ingredients:
         parsed_ingredient_name = ingredients['ingredient_name'].lower().strip()
         parsed_quantity_unit = ingredients['quantity_unit'].lower().strip()
         if(len(parsed_quantity_unit) > 1 and parsed_quantity_unit[-1] == 's' ):
             parsed_quantity_unit = parsed_quantity_unit[:-1]
-        print('ingredient unit afetr parse', parsed_quantity_unit)
         ingredient = recipe.ingredients.create(ingredient_name=parsed_ingredient_name,
                         ingredient_quantity=str(round(float(ingredients['ingredient_quantity']),2)), quantity_unit=parsed_quantity_unit, recipe=recipe)
-        print(ingredient)
         ingredient.save()
